+++ io/args.go
@@ -18,7 +18,7 @@ func ReadArgs(instances []types.Instance) []types.Instance {
 		if inst.DBMS == -1 {
 			errors.ThrowArgError(os.Args)
 		}
-		inst.DSN = os.Args[2]
+		inst.DSN = []byte(os.Args[2])
 
 		if len(os.Args) == 4 {
 			inst.Dbname = os.Args[3]