+++ ui/statemachine.go
@@ -12,7 +12,7 @@ var (
 	State     types.State_t
 	Laststate types.State_t
 	Interval  time.Duration = 500 * time.Millisecond
-	Ticker    *time.Ticker  = time.NewTicker(1 * time.Second)
+	PoolNdx   int           = 0
 )
 
 func InterfaceLoop(instances []types.Instance, cpool []*sql.DB) {
@@ -47,28 +47,48 @@ func InterfaceLoop(instances []types.Instance, cpool []*sql.DB) {
 		case types.PROCESSLIST:
 			DisplayProcesslist(t, cpool)
 			Laststate = types.PROCESSLIST
+			if State == types.PROCESSLIST {
+				State = types.MENU
+			}
 			break
 		case types.DB_DASHBOARD:
 			Laststate = types.DB_DASHBOARD
+			DisplayDbDashboard(t, cpool)
+			if State == types.DB_DASHBOARD {
+				State = types.MENU
+			}
 			break
 		case types.MEM_DASHBOARD:
 			Laststate = types.MEM_DASHBOARD
+			DisplayMemory(t, cpool)
+			if State == types.MEM_DASHBOARD {
+				State = types.MENU
+			}
 			break
 		case types.ERR_LOG:
+			DisplayErrorLog(t, cpool)
 			Laststate = types.ERR_LOG
+			if State == types.ERR_LOG {
+				State = types.MENU
+			}
 			break
 		case types.LOCK_LOG:
-			/*
-				Assuming this shows current locks in backend
-			*/
+			DisplayLocks(t, cpool)
 			Laststate = types.LOCK_LOG
+			if State == types.LOCK_LOG {
+				State = types.MENU
+			}
 			break
 		case types.CONFIGS:
 			/*
 				Force user to this state if no configs are found and if launched w/o args.
 				Prompt user to connect to database
 			*/
+			DisplayConfigs(t, instances, cpool)
 			Laststate = types.CONFIGS
+			if State == types.CONFIGS {
+				State = types.MENU
+			}
 			break
 		case types.HELP:
 			/*
@@ -76,13 +96,15 @@ func InterfaceLoop(instances []types.Instance, cpool []*sql.DB) {
 			*/
 			DrawHelp(t)
 			Laststate = types.HELP
+			if State == types.HELP {
+				State = types.MENU
+			}
 			break
 		case types.QUIT:
 			/*
 				Perform cleanup and close program
 			*/
 			//t.Close() fixes input buffer overflow ?
-			Ticker.Stop()
 			for _, conn := range cpool {
 				conn.Close()
 			}